requirements.txt
appdirs==1.4.4
astroid==2.4.2
beautifulsoup4==4.9.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
distlib==0.3.1
filelock==3.0.12
google-api-core==1.22.0
google-auth==1.19.2
googleapis-common-protos==1.52.0
grpcio==1.30.0
httplib2==0.18.1
idna==2.10
imutils==0.5.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.0
opencv-python==4.3.0.36
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.5.3
pytz==2020.1
requests==2.24.0
rsa==4.6
scipy==1.5.1
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.9
virtualenv==20.0.26
wrapt==1.12.1
google-cloud-videointelligence
pandas
openpyxl
xlrd